requirements.txt
pyautogui
requests
Pillow
packaging
requests
psutil
pyarmor
pyinstaller